requirements.txt
ipykernel
psycopg[binary]
Faker
django
flake8
python-dotenv
django-split-settings
django-debug-toolbar
pytest
pytest-cov
